+++ package.json
@@ -49,6 +49,9 @@
     "html-webpack-plugin": "^2.24.1",
     "node-sass": "^4.5.3",
     "nodemon": "^1.11.0",
+    "postcss": "^6.0.14",
+    "postcss-cssnext": "^3.0.2",
+    "postcss-import": "^11.0.0",
     "postcss-loader": "^2.0.8",
     "prepush": "^3.1.11",
     "react-hot-loader": "^3.1.1",
