+++ postcss.config.js
@@ -1,0 +1,10 @@
+module.exports = {
+  plugins: {
+    'postcss-import': {
+      path: 'source/css/',
+    },
+    'postcss-cssnext': {
+      browsers: ['last 2 versions', '> 5%'],
+    },
+  },
+};
